+++ R/enum.R
@@ -17,7 +17,8 @@
   gamma_mix = 13,
   lognormal_mix = 14,
   nbinom2_mix = 15,
-  gengamma = 16
+  gengamma = 16,
+  betabinomial = 17
 )
 .valid_link <- c(
   identity = 0,